moved zarrdescriber to subpackage
